jira_integration/requirements.txt
configparser
jsons
time
sys
jira
atlassian-python-api
csv
numpy
pandas
requests